% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_cc}
\alias{ex_cc}
\title{Simulated case cohort data}
\format{A data frame with 1571 rows and 7 variables:
\describe{
  \item{t}{Time to event or censoring}
  \item{d}{Indicator of whether event 1 occurred (d=1), or not (d=0)}
  \item{x}{Partially observed continuous covariate}
  \item{z}{Fully observed covariate}
  \item{in.subco}{A binary indicator of whether the subject is in the sub-cohort}
  \item{id}{An id variable}
  \item{entertime}{The entry time variable to be used in the analysis}
}}
\usage{
ex_cc
}
\description{
A dataset containing simulated case cohort data, where the sub-cohort was a 10\% random sample of the full cohort.
}
\keyword{datasets}
