% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_coarsening}
\alias{ex_coarsening}
\title{Simulated example data with a coarsened factor covariate}
\format{
A data frame with 100 rows and 4 variables:
\describe{
  \item{x}{Factor variable with 3 levels (a,b,c) and some missing values}
  \item{xobs}{Variable which contains coarsening information for x}
  \item{z}{Fully observed continuous covariate}
  \item{z}{Fully observed continuous outcome}
}
}
\usage{
ex_coarsening
}
\description{
A simulated dataset with a factor level covariate subject to coarsening.
}
\keyword{datasets}
