% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smcfcs.r
\docType{data}
\name{ex_coverr}
\alias{ex_coverr}
\title{Simulated example data with binary outcome and covariate measured with error}
\format{A data frame with 10000 rows and 4 variables:
\describe{
  \item{y}{Binary outcome}
  \item{x}{An entirely missing covariate, representing the true covariate which is measured with
  error}
  \item{w1}{The first error-prone measurement of x, available on all 10000}
  \item{w2}{The second error-prone measurement of x, available on a random 10\%}
}}
\usage{
ex_coverr
}
\description{
A dataset containing simulated data where the binary outcome depends linearly
on a covariate that is not directly measured. Instead, it is measured with error
by two error prone measurements, w1 and w2.
}
\keyword{datasets}

