% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_coxquad}
\alias{ex_coxquad}
\title{Simulated example data with time to event outcome and quadratic covariate effects}
\format{A data frame with 1000 rows and 6 variables:
\describe{
  \item{t}{Time to event or censoring}
  \item{d}{Binary indicator of whether event occurred or individual was censored}
  \item{z}{Fully observed covariate, with linear effect on outcome (on log hazard scale)}
  \item{x}{Partially observed normally distributed covariate, with quadratic effect on outcome (on log hazard scale)}
  \item{xsq}{The square of x, which thus has missing values also}
  \item{v}{An auxiliary variable (i.e. not contained in the substantive model)}
}}
\usage{
ex_coxquad
}
\description{
A dataset containing simulated data where a time to event outcome depends quadratically
on a partially observed covariate.
}
\keyword{datasets}
