% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_dtsam}
\alias{ex_dtsam}
\title{Simulated discrete time survival data set}
\format{
A data frame with 1000 rows and 8 variables:
\describe{
  \item{x1}{A binary variable with missing values}
  \item{x2}{A fully observed continuous variable}
  \item{failtime}{The discrete failure/censoring time}
  \item{d}{Indicator of failure (=1) or censoring (=0)}
}
}
\usage{
ex_dtsam
}
\description{
A dataset containing simulated discrete time survival data.
}
\keyword{datasets}
