% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_flexsurv}
\alias{ex_flexsurv}
\title{Simulated example data with time-to-event Weibull outcome and two covariates}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{t}{Time to event (d=1) or censoring (d=0)}
  \item{d}{Event indicator}
  \item{x}{Partially observed binary covariate}
  \item{z}{Fully observed continuous covariate}
}
}
\usage{
ex_flexsurv
}
\description{
A dataset containing simulated data where the time-to-event outcome is Weibull
distributed with two covariates, one of which is partially observed.
}
\keyword{datasets}
