% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_lininter}
\alias{ex_lininter}
\title{Simulated example data with continuous outcome and interaction between two partially observed covariates}
\format{A data frame with 1000 rows and 4 variables:
\describe{
  \item{y}{Continuous outcome}
  \item{x1}{Partially observed normally distributed covariate}
  \item{x2}{Partially observed binary covariate}
}}
\usage{
ex_lininter
}
\description{
A dataset containing simulated data where the outcome depends on both main
effects and interaction of two partially observed covariates.
}
\keyword{datasets}
