% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_linquad}
\alias{ex_linquad}
\title{Simulated example data with continuous outcome and quadratic covariate effects}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{y}{Continuous outcome}
  \item{z}{Fully observed covariate, with linear effect on outcome}
  \item{x}{Partially observed normally distributed covariate, with quadratic effect on outcome}
  \item{v}{An auxiliary variable (i.e. not contained in the substantive model)}
}
}
\usage{
ex_linquad
}
\description{
A dataset containing simulated data where the outcome depends quadratically
on a partially observed covariate.
}
\keyword{datasets}
