% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_logisticquad}
\alias{ex_logisticquad}
\title{Simulated example data with binary outcome and quadratic covariate effects}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{y}{Binary outcome}
  \item{z}{Fully observed covariate, with linear effect on outcome (on log odds scale)}
  \item{x}{Partially observed normally distributed covariate, with quadratic effect on outcome (on log odds scale)}
  \item{v}{An auxiliary variable (i.e. not contained in the substantive model)}
}
}
\usage{
ex_logisticquad
}
\description{
A dataset containing simulated data where the binary outcome depends quadratically
on a partially observed covariate.
}
\keyword{datasets}
