% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_ncc}
\alias{ex_ncc}
\title{Simulated nested case-control data}
\format{
A data frame with 728 rows and 8 variables:
\describe{
  \item{t}{Time to event or censoring}
  \item{d}{Indicator of whether event 1 occurred (d=1), or not (d=0)}
  \item{x}{Partially observed binary covariate}
  \item{z}{Fully observed covariate}
  \item{id}{An id variable}
  \item{numrisk}{Number of patients at risk at time of case's event}
  \item{setno}{The case-control set number}
  \item{case}{Binary indicator of case (=1) or control (=0)}
}

A data frame with 1000 rows and 8 variables:
\describe{
  \item{x1}{A binary variable with missing values}
  \item{x2}{A fully observed continuous variable}
  \item{failtime}{The discrete failure/censoring time}
  \item{d}{Indicator of failure (=1) or censoring (=0)}
}
}
\usage{
ex_ncc

ex_ncc
}
\description{
A dataset containing simulated nested case-control data.

A dataset containing simulated discrete time survival data.
}
\keyword{datasets}
