% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_poisson}
\alias{ex_poisson}
\title{Simulated example data with count outcome, modelled using Poisson regression}
\format{A data frame with 1000 rows and 3 variables:
\describe{
  \item{y}{Count outcome}
  \item{z}{Fully observed covariate, with linear effect on outcome}
  \item{x}{Partially observed normally distributed covariate, with linear effect on outcome}
}}
\usage{
ex_poisson
}
\description{
A dataset containing simulated data where the count outcome depends on two
covariates, x and z, with missing values in x. The substantive model is
Poisson regression.
}
\keyword{datasets}
