% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence.R
\name{plot.smcfcs}
\alias{plot.smcfcs}
\title{Plot convergence of a smcfcs object}
\usage{
\method{plot}{smcfcs}(x, ...)
}
\arguments{
\item{x}{An object of class 'smcfcs'}

\item{...}{Additional parameters to pass on to ggplot2::facet_wrap(),
eg. nrow = 2}
}
\value{
A ggplot2 object, containing the convergence plots, facetted per
covariate in the substantive model
}
\description{
Visualises the contents of smCoefIter.
}
\examples{
\dontrun{
# Use simulated competing risks example in package
imps <- smcfcs(
originaldata = ex_compet,
smtype = "compet",
smformula = list(
"Surv(t, d == 1) ~ x1 + x2",
"Surv(t, d == 2) ~ x1 + x2"
),
method = c("", "", "norm", "norm")
)

plot(imps)
}

}
