% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{smcfcs.compet}
\alias{smcfcs.compet}
\title{Competing risks model compatible fully conditional specification
multiple imputation of covariates.}
\usage{
smcfcs.compet(originaldata, timevar, causevar, linpred, method,
  predictorMatrix = NULL, m = 5, maxit = 10, rjlimit = 1000,
  noisy = NULL)
}
\arguments{
\item{originaldata}{Original data frame}

\item{method}{A vector of strings specifying what type of regression model
should be used to impute each variable}
}
\description{
Multiply imputes missing covariate values using substantive model compatible
fully conditional specification multiple imputation for competing risks
data, assuming Cox models for the cause specific hazard functions.
}

