% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{smcfcs.coxph}
\alias{smcfcs.coxph}
\title{Cox proportional hazards model compatible fully conditional specification
multiple imputation of covariates.}
\usage{
smcfcs.coxph(originaldata, smformula, method, predictorMatrix = NULL, m = 5,
  numit = 10, rjlimit = 1000, noisy = FALSE)
}
\arguments{
\item{originaldata}{The original data frame with missing values.}

\item{smformula}{The Cox substantive model formula, with the left hand side
of the formula of the form "Surv(t,delta)", where t specifies the event time
variable and delta specifies the binary event indicator variable.}

\item{method}{A required vector of strings specifying for each variable either
that it does not need to be imputed (""), the type of regression model to be
be used to impute (possible values are "norm", "logreg", "poisson", "podds"
and "mlogit") or an expression to passively impute the variable. See the
examples.}

\item{predictorMatrix}{An optional predictor matrix. If specified, the matrix defines which
covariates will be used as predictors in the imputation models
(the outcome must not be included). The i'th row of the matrix should consist of
0s and 1s, with a 1 in the j'th column indicating the j'th variable be used
as a covariate when imputing the i'th variable. If not specified, when
imputing a given variable, the imputation model covariates are the other
partially observed variables and any fully observed variables (if present).
Note that the outcome variable is implicitly conditioned on by the rejection
sampling scheme used by smcfcs.}

\item{m}{The number of imputed datasets to generate.}

\item{numit}{The number of iterations to run when generating each imputation.}

\item{rjlimit}{Specifies the maximum number of attempts which should be made
when using rejection sampling to draw from imputation models.}
}
\description{
Multiply imputes missing covariate values using substantive model compatible
fully conditional specification multiple imputation for a Cox proportional
hazards substantive model.
}

