% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{smcfcs.lm}
\alias{smcfcs.lm}
\title{Linear regression substantive model compatible fully conditional specificiation
multiple imputation of covariates.}
\usage{
smcfcs.lm(originaldata, smformula, method, predictorMatrix = NULL, m = 5,
  maxit = 10, rjlimit = 1000, noisy = NULL)
}
\arguments{
\item{originaldata}{Original data frame}
}
\value{
a list of data frames containing the multiply imputed datasets.
}
\description{
Multiple imputes missing covariates values using substantive model compatible
fully conditional specification multiple imputation for linear regression
substantive models.
}

