% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{smcfcs.logistic}
\alias{smcfcs.logistic}
\title{Logistic regression substantive model compatible fully conditional specification
multiple imputation of covariates.}
\usage{
smcfcs.logistic(originaldata, smformula, method, predictorMatrix = NULL,
  m = 5, maxit = 10, rjlimit = 1000, noisy = NULL)
}
\arguments{
\item{originaldata}{Original data frame}

\item{smformula}{The formula of the substantive model}

\item{method}{A vector of strings specifying what type of regression model
should be used to impute each variable}
}
\description{
Multiple imputes missing covariates values using substantive model compatible
fully conditional specification multiple imputation for logistic regression
substantive models.
}

