% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{smcfcs.nestedcc}
\alias{smcfcs.nestedcc}
\title{Substantive model compatible fully conditional specification imputation of covariates for nested case control
studies}
\usage{
smcfcs.nestedcc(originaldata, smformula, set, substudy, cc, id, method,
  predictorMatrix = NULL, m = 5, numit = 10, rjlimit = 1000,
  noisy = FALSE)
}
\arguments{
\item{originaldata}{The original data frame with missing values.}

\item{smformula}{The formula of the substantive model. For \code{"coxph"} substantive
models the left hand side should be of the form \code{"Surv(t,d)"}. For \code{"compet"}
substantive models, a list should be passed consisting of the Cox models
for each cause of failure (see example).}

\item{set}{variable identifying matched sets in nested case-control study}

\item{substudy}{variable which is an indicator of who is in the sub-study}

\item{method}{A required vector of strings specifying for each variable either
that it does not need to be imputed (""), the type of regression model to be
be used to impute. Possible values are \code{"norm"} (normal linear regression),
\code{"logreg"} (logistic regression), \code{"poisson"} (Poisson regression),
\code{"podds"} (proportional odds regression for ordered categorical variables),
\code{"mlogit"} (multinomial logistic regression for unordered categorical variables),
or a custom expression which defines a passively imputed variable, e.g.
\code{"x^2"} or \code{"x1*x2"}.}

\item{predictorMatrix}{An optional predictor matrix. If specified, the matrix defines which
covariates will be used as predictors in the imputation models
(the outcome must not be included). The i'th row of the matrix should consist of
0s and 1s, with a 1 in the j'th column indicating the j'th variable be used
as a covariate when imputing the i'th variable. If not specified, when
imputing a given variable, the imputation model covariates are the other
covariates of the substantive model which are partially observed
(but which are not passively imputed) and any fully observed covariates (if present)
in the substantive model. Note that the outcome variable is implicitly conditioned
on by the rejection sampling scheme used by smcfcs, and should not be specified as a predictor
in the predictor matrix.}

\item{m}{The number of imputed datasets to generate. The default is 5.}

\item{numit}{The number of iterations to run when generating each imputation.
In a (limited) range of simulations good performance was obtained with the
default of 10 iterations. However, particularly when the proportion of missingness
is large, more iterations may be required for convergence to stationarity.}

\item{rjlimit}{Specifies the maximum number of attempts which should be made
when using rejection sampling to draw from imputation models. If the limit is reached
when running a warning will be issued. In this case it is probably advisable to
increase the \code{rjlimit} until the warning does not appear.}

\item{noisy}{logical value (default FALSE) indicating whether output should be noisy, which can
be useful for debugging or checking that models being used are as desired.}

\item{case}{variable which indicates who is a case/control in the nested case-control sample. Note that this is distinct from d.}
}
\description{
Multiply imputes missing covariate values using substantive model compatible
fully conditional specification for nested case control studies.
}
\details{
This version of \code{smcfcs} is designed for use with nested case control studies. The function's arguments are the same
as for the main smcfcs function, except for \code{smformula}, \code{in.subco}, and \code{sampfrac} - see above
for details on how these should be specified.
}
\author{
Ruth Keogh \email{ruth.keogh@lshtm.ac.uk}

Jonathan Bartlett \email{jwb133@googlemail.com}
}

