% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ex_ncc}
\alias{ex_ncc}
\title{Simulated nested case-control data}
\format{A data frame with 1571 rows and 7 variables:
\describe{
  \item{t}{Time to event or censoring}
  \item{d}{Indicator of whether event 1 occurred (d=1), or not (d=0)}
  \item{x}{Partially observed binary covariate}
  \item{z}{Fully observed covariate}
  \item{id}{An id variable}
  \item{numrisk}{Number of patients at risk at time of case's event}
  \item{setno}{The case-control set number}
  \item{case}{Binary indicator of case (=1) or control (=0)}
}}
\usage{
ex_ncc
}
\description{
A dataset containing simulated nested case-control data.
}
\keyword{datasets}
