% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{mips}
\alias{mips}
\title{Multiple imputation for confounders in propensity score analysis.}
\usage{
mips(originaldata, omtype, omformula, method, predictorMatrix = NULL, m = 5,
  numit = 10, rjlimit = 1000, noisy = FALSE)
}
\arguments{
\item{omtype}{A string specifying the type of outcome model. Possible
values are \code{"lm"}, \code{"logistic"}, \code{"poisson"}, \code{"coxph"}
 and \code{"compet"}.}

\item{omformula}{The formula of the outcome model. For \code{"coxph"} outcome
models the left hand side should be of the form \code{"Surv(t,d)"}. For \code{"compet"}
outcome models, a list should be passed consisting of the Cox models
for each cause of failure.}

\item{psformula}{The formula for the propensity score model.

\code{mips} imputes missing values in confounders prior to propensity score analysis.}
}
\description{
Multiply imputes missing covariate values using substantive model compatible
fully conditional specification.
}

