% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mips.r
\name{mips}
\alias{mips}
\title{Multiple imputation for confounders in propensity score analysis.}
\usage{
mips(originaldata, omtype, omformula, method, predictorMatrix = NULL, m = 5,
  numit = 10, rjlimit = 1000, noisy = FALSE, psformula)
}
\arguments{
\item{originaldata}{The original data frame with missing values.}

\item{omtype}{A string specifying the type of outcome model. Possible
values are \code{"lm"}, \code{"logistic"}, \code{"poisson"}, \code{"coxph"}
 and \code{"compet"}.}

\item{omformula}{The formula of the outcome model. For \code{"coxph"} outcome
models the left hand side should be of the form \code{"Surv(t,d)"}. For \code{"compet"}
outcome models, a list should be passed consisting of the Cox models
for each cause of failure.}

\item{method}{A required vector of strings specifying for each variable either
that it does not need to be imputed (""), the type of regression model to be
be used to impute. Possible values are \code{"norm"} (normal linear regression),
\code{"logreg"} (logistic regression), \code{"poisson"} (Poisson regression),
\code{"podds"} (proportional odds regression for ordered categorical variables),
\code{"mlogit"} (multinomial logistic regression for unordered categorical variables),
or a custom expression which defines a passively imputed variable, e.g.
\code{"x^2"} or \code{"x1*x2"}.}

\item{predictorMatrix}{An optional predictor matrix. If specified, the matrix defines which
covariates will be used as predictors in the imputation models
(the outcome must not be included). The i'th row of the matrix should consist of
0s and 1s, with a 1 in the j'th column indicating the j'th variable be used
as a covariate when imputing the i'th variable. If not specified, when
imputing a given variable, the imputation model covariates are the other
covariates of the substantive/outcome model which are partially observed
(but which are not passively imputed) and any fully observed covariates (if present)
in the substantive/outcome model. Note that the outcome variable is implicitly conditioned
on by the rejection sampling scheme used by smcfcs, and should not be specified as a predictor
in the predictor matrix.}

\item{m}{The number of imputed datasets to generate. The default is 5.}

\item{numit}{The number of iterations to run when generating each imputation.
In a (limited) range of simulations good performance was obtained with the
default of 10 iterations. However, particularly when the proportion of missingness
is large, more iterations may be required for convergence to stationarity.}

\item{rjlimit}{Specifies the maximum number of attempts which should be made
when using rejection sampling to draw from imputation models. If the limit is reached
when running a warning will be issued. In this case it is probably advisable to
increase the \code{rjlimit} until the warning does not appear.}

\item{noisy}{logical value (default FALSE) indicating whether output should be noisy, which can
be useful for debugging or checking that models being used are as desired.}

\item{psformula}{The model formula for the propensity score model.}
}
\value{
A list containing the imputed datasets.
}
\description{
Multiply imputes missing covariate values using an adaptation of the substantive
model compatible fully conditional specification (SMC-FCS) approach to multiple imputation.
}
\details{
\code{mips} imputes missing values in confounders prior to propensity score analysis.
}
\examples{
#set random number seed to make results reproducible
set.seed(123)

#simulate a dataset
n <- 200
x <- rnorm(n)
psxb <- x
pspr <- exp(psxb)/(1+exp(psxb))
z <- 1*(runif(n)<pspr)
y <- x + z + rnorm(n)
data <- data.frame(x,z,y)

#make some x values missing completely at random
data$x[runif(n)<0.5] <- NA

#generate multiple imputations using mips
imps <- mips(data, omtype="lm", omformula="y~x+z",
             method=c("norm","",""),psformula="z~x",m=5)
#now do some propensity score analysis using the imputed datasets
}

