% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{smcfcs.compet}
\alias{smcfcs.compet}
\title{Competing risks model compatible fully conditional specification
multiple imputation of covariates.}
\usage{
smcfcs.compet(originaldata, timevar, causevar, linpred, method,
  predictorMatrix = NULL, m = 5, maxit = 10, rjlimit = 1000,
  noisy = NULL)
}
\arguments{
\item{originaldata}{Original data frame}

\item{method}{A vector of strings specifying what type of regression model
should be used to impute each variable}

\item{predictorMatrix}{An optional predictor matrix. If specified, the matrix defines which
variables will be used as predictors in the imputation models
(not including the outcome). The i'th row of the matrix should consist of
0s and 1s, with a 1 in the j'th column indicating the j'th variable be used
as a covariate when imputing the i'th variable.}
}
\description{
Multiply imputes missing covariate values using substantive model compatible
fully conditional specification multiple imputation for competing risks
data, assuming Cox models for the cause specific hazard functions.
}

