% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smcfcs.r
\name{smcfcs.lm}
\alias{smcfcs.lm}
\title{Linear regression substantive model compatible fully conditional specification
multiple imputation of covariates.}
\usage{
smcfcs.lm(originaldata, smformula, method, predictorMatrix = NULL, m = 5,
  maxit = 10, rjlimit = 1000, noisy = NULL)
}
\arguments{
\item{originaldata}{Original data frame}

\item{smformula}{The formula of the substantive model}

\item{method}{A vector of strings specifying what type of regression model
should be used to impute each variable}

\item{predictorMatrix}{An optional predictor matrix. If specified, the matrix defines which
variables will be used as predictors in the imputation models
(not including the outcome). The i'th row of the matrix should consist of
0s and 1s, with a 1 in the j'th column indicating the j'th variable be used
as a covariate when imputing the i'th variable.}
}
\value{
a list of data frames containing the multiply imputed datasets.
}
\description{
Multiple imputes missing covariate values using substantive model compatible
fully conditional specification multiple imputation for linear regression
substantive models.
}

